Reduce CPU load
Lower target framerate from 100 fps to 30 fps again.
Adapt script parameters to the lowered framerate.
Reduction of the target framerate does not harm input latency
anymore, since we now wait on the keyboard input thread inside
the main loop. This has the effect that the target framerate
will be raised dynamically on keyboard input "as needed".