Disable waiting for mouse move event-handler
Do not wait for completion of the mouse move event handler. This should
reduce input lag, but will slightly increase CPU load.
Should result in a much smoother and less jerky performance of the
effects (on mouse movement). Note that we still have to wait for the
completion of keyboard events, so that we are able to inhibit them if
requested, e.g. by the macro subsystem.