Add a failsafe mode to Eruption
Do not crash the driver, but switch to the newly introduced failsafe
mode as soon as an irrecoverable error is encountered:

* Syntax errors in Lua scripts
* Runtime errors in Lua scripts
* Corrupted .profile files

In failsafe mode, the LEDs will just stay dark, only the minimal subset
of features are supported like e.g. the keyboard shortcuts to switch to
a different slot.

References #54