Try to avoid blocking the main/input threads
The intention here is to reduce jitter and to avoid holding any locks
that the input thread requires.
Hangs may get triggered on some systems when the state of the audio hardware
changes, and subsequently the eruption-audio-proxy daemon reports a
notification.

Use unbounded channels for rx/tx of audio samples.
Use unbounded channels for rx/tx to the device I/O thread.

References #140