Introduce a delay after each step of initialization
Introduce a short delay after each step of the initialization sequence.
This is a try to fix a problem with failing device initialization,
seen on some KVM switches.

This patch shortens the delay somewhat and places it in front of the
status polling loop. This is done to avoid introducing lag due to excessive
sleeping.

References #161