Use order-preserving containers
Use an `IndexMap` to store the associated manifests of a profile, since we
actually care about the order of the script files, to achieve correct
rendering results.

Switch to using `BTreeMaps` instead of `HashMaps` for storing
parameters for the same reason. Correctness would be maintained even
when using `HashMaps` though.