Add support for GNOME 45 on Wayland
GNOME 45 on Wayland does not support to query which window is currently the
focused one. The `gnome_shellext` sensor backend can make use of a
GNOME shell extension that tracks the currently active window and sends events
over a FIFO for us to process.

NOTE:
For this to work, the `Eruption Sensor` GNOME shell extension is required to
be running!